add check for folder existance

+++ tools/install_fadtk.sh
@@ -1,7 +1,8 @@
 #/bin/bash
 
-
-rm -rf fadtk
+if [ -d "fadtk" ]; then
+    rm -rf fadtk
+fi
 
 # NOTE(jiatong): a versa-specialized implementation for fadtk
 git clone https://github.com/ftshijt/fadtk.git
