+++ tools/install_fairseq.sh
@@ -1,7 +1,8 @@
 #/bin/bash
 
-
-rm -rf fairseq
+if [ -d "fairseq" ]; then
+    rm -rf fairseq
+fi
 
 git clone -b versa https://github.com/ftshijt/fairseq.git
 cd fairseq
