+++ tools/install_warpq.sh
@@ -1,7 +1,8 @@
 #/bin/bash
 
-
-rm -rf WARP-Q
+if [ -d "WARP-Q" ]; then
+    rm -rf WARP-Q
+fi
 
 # # NOTE(jiatong): a versa-specialized implementation for WARP-Q
 git clone https://github.com/ftshijt/WARP-Q.git
